apps/api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
